Moved tests to appropriate files
